update 102- 103- 104-

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int fibonacci(int n)
+long int fibonacci(int n)
 {
 	if (n == 1)
 	{ 
@@ -19,18 +19,18 @@ int fibonacci(int n)
 int main(void)
 {
 	int i = 1;
-	int fib;
+	long int fib;
 
 	while (i <= 10)
 	{
 		fib = fibonacci(i);
 		if (i == 10)
 		{
-			printf("%d", fib);
+			printf("%ld", fib);
 		}
 		else
 		{
-			printf("%d, ", fib);
+			printf("%ld, ", fib);
 		}
 		i++;
 	}
